Extract collection/franchise template-var normalizers into modules/output_collections.py
Sixth refactor pass at modules/output.py.  Moved the collection-shaped
template variable machinery -- the collection-side counterpart to what
PR #1448 did for playlists.

Constant moved:
  * FRANCHISE_DYNAMIC_CHILD_FIELD_SPECS  (17-entry dict of override->kind)

Functions moved (from output.py to output_collections.py):
  * _parse_tmdb_person_window            (69 lines: JSON/py-literal/CSV kv)
  * _normalize_collection_template_var_value  (dispatches by field name)
  * _normalize_dynamic_child_override_value   (dispatches by kind)
  * _expand_franchise_dynamic_child_overrides (walk + flatten)
  * _normalize_settings_section_value    (top-level ignore_ids/imdb_ids)

Bonus cleanup: removed 'import ast' from output.py -- no longer used.

Dependencies: modules.output_values only.

Public surface: output.py re-imports the six names with '# noqa: F401'.

Impact:
- modules/output.py: 3119 -> 2971 (-148 / -4.7%)   -- BROKE BELOW 3000!
- modules/output_collections.py: 226 lines new
- Zero behavior change

Cumulative on output.py: 3833 -> 2971 (-862 / -22.5%)

All 1077 tests pass. Ruff + black clean.
